fix(narrated): plan voiceover stories reliably, make Clean up speech real, recover from planning failures (#1166)
* fix(narrated): plan voiceover stories reliably, make Clean up speech real, recover from planning failures

Narrated stories on the iPhone failed at planning after the KRI-133
semantic planner shipped: the model copied every voiceover line into
text_bindings (a fast-montage-only lane that guided stories never draw),
and a photo reused across two chapters with different lines raised a
terminal `conflicting_text_bindings` in the scheduler. The same input
failed every time, so "Try it again" could never succeed, and each
failure also spent a render attempt.

Planning
- Drop text bindings before validation for every direction except
  fast_montage; the scheduler ignores them there too. Creator copy that
  only lived in a binding moves to a chapter thought (bound shots get
  their own chapter); copy that provably cannot be placed renders what
  fits with an `unplaceable_creator_caption` repair, matching main.
- Narrated stories: AI draft thoughts are blanked before validation so
  only the title and voiceover captions render ("No other text besides
  the title"); a spoken quote counts as script unless it is placed.
- Invented surplus media aliases (m000) are dropped; repeated sources
  merge; retry hints for unknown media and dropped captions.
- Fix NameError in EditProposalSnapshot.validate_beat_media when a
  story snapshot carries montage_audio (also on main).

Clean up speech (GUIDED_VOICEOVER_SPEECH_CLEANUP_ENABLED, default off)
- On a Clean choice the draft task cuts the consented CutPlan into a
  content-addressed WAV under users/{owner}/plan/{item}/speech-cleanup/
  and plans, captions and renders (phone + cloud) against it, with
  provenance on NarrationTrack.speech_cleanup bound to the Job's
  preflight snapshot. Previously the choice was recorded but the raw
  voiceover shipped. Off: Clean is refused before an attempt is used.

Failures the creator can act on
- Retryable planning failures refund the render attempt (max 2 per
  plan); the failure event and projection carry the real reason,
  retryable/can_retry and the attempt budget; the current iOS card shows
  the reason instead of "Your clips are ready.".
- A message after a failed session keeps the creator's brief.
- The draft task retries transient DB disconnects and resumes its own
  drafting step. Admin plan-item debug accepts fractional durations.


* test(ios): simulator proof that a cleaned guided voiceover renders on the iPhone

phone-photo-render-e2e.py gains a narrated_story_cleaned case: a 48 s
recording with 5.6 s pauses goes through the real speech-cleanup
analysis and the real draft-time derivative builder
(build_cleaned_narration) into a 16.45 s WAV, compiled with
compile_phone_guided_plan into a pilot-valid recipe with one narration
clip. testCleanedNarratedStoryRendersOnTheIPhone renders it through the
production device resolver/exporter and asserts the cleaned duration,
audible narration, caption timing at the cleaned word times, and that no
removed pause is audible (fails with the raw recording: 3.27 s silence).


---------

+++ src/apps/api/app/routes/admin_plan_items.py
@@ -130,7 +130,9 @@ class ProposalBriefPayload(BaseModel):
     goal_length: int
     creator_request_length: int
     pace: str
-    duration_s: int
+    # ProposalDuration is int | float: a narrated brief carries the
+    # voiceover's fractional length, which an int field rejects (500).
+    duration_s: float
 
 
 class ProposalFailurePayload(BaseModel):
@@ -160,7 +162,7 @@ class LastApprovedSummaryPayload(BaseModel):
 
 class DraftSummaryPayload(BaseModel):
     beat_count: int
-    duration_s: int
+    duration_s: float
 
 
 class PlannerFallbackPayload(BaseModel):
